requirements.txt
absl-py==2.2.0
altgraph==0.17.4
attrs==25.3.0
cffi==1.17.1
contourpy==1.3.0
cv2_enumerate_cameras==1.1.18.3
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.56.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
kiwisolver==1.4.7
matplotlib==3.9.4
mediapipe==0.10.9
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
protobuf==3.20.3
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-osc==1.8.3
pywin32-ctypes==0.2.3
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
TKinterModernThemes==1.10.4
zipp==3.21.0
